Flyttet logikken for å sjekke om det er endret utbetaling inn til AndelTilkjentYtelse.kt
